requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.7
